+++ _tests/test/common/forms/directives_test.dart
@@ -357,5 +357,4 @@ void main() {
   });
 }
 
-@proxy
 class MockValueAccessor extends Mock implements ControlValueAccessor {}